+++ cursus/libft/ft_strlen.c
@@ -9,10 +9,11 @@
 /*   Updated: 2022/12/01 21:16:30 by daewoole         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
+#include "libft.h"
 
-int	ft_strlen(char *str)
+size_t	ft_strlen(char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (*(str + i) != '\0')
